CRDCDH-1861 Move dbGaPID to Section C
