+++ src/content/questionnaire/sections/B.tsx
@@ -25,7 +25,6 @@ import AddRemoveButton from "../../../components/AddRemoveButton";
 import PlannedPublication from "../../../components/Questionnaire/PlannedPublication";
 import { InitialQuestionnaire } from "../../../config/InitialValues";
 import TransitionGroupWrapper from "../../../components/Questionnaire/TransitionGroupWrapper";
-import SwitchInput from "../../../components/Questionnaire/SwitchInput";
 import useFormMode from "../../../hooks/useFormMode";
 import FundingAgency from "../../../components/Questionnaire/FundingAgency";
 import SelectInput from "../../../components/Questionnaire/SelectInput";
@@ -80,10 +79,6 @@ const FormSectionB: FC<FormSectionProps> = ({ SectionOption, refs }: FormSection
   const [fundings, setFundings] = useState<KeyedFunding[]>(
     data.study?.funding?.map(mapObjectWithKey) || []
   );
-  const [isDbGapRegistered, setIsdbGaPRegistered] = useState<boolean>(
-    data.study?.isDbGapRegistered
-  );
-  const [dbGaPPPHSNumber, setDbGaPPPHSNumber] = useState<string>(data.study?.dbGaPPPHSNumber);
 
   const programKeyRef = useRef(new Date().getTime());
   const formContainerRef = useRef<HTMLDivElement>();
@@ -111,9 +106,6 @@ const FormSectionB: FC<FormSectionProps> = ({ SectionOption, refs }: FormSection
     if (!formObject.study) {
       combinedData.study = InitialQuestionnaire.study;
     }
-    if (!formObject?.study?.dbGaPPPHSNumber) {
-      combinedData.study.dbGaPPPHSNumber = "";
-    }
 
     // Reset publications if the user has not entered any publications
     if (!formObject.study.publications || formObject.study.publications.length === 0) {
@@ -134,6 +126,13 @@ const FormSectionB: FC<FormSectionProps> = ({ SectionOption, refs }: FormSection
           ? plannedPublication.expectedDate
           : "",
       })) || [];
+
+    combinedData.study = {
+      // Restore dbGaPID from Section C
+      isDbGapRegistered: data?.study?.isDbGapRegistered,
+      dbGaPPPHSNumber: data?.study?.dbGaPPPHSNumber,
+      ...combinedData.study,
+    };
 
     return { ref: formRef, data: combinedData };
   };
@@ -187,13 +186,6 @@ const FormSectionB: FC<FormSectionProps> = ({ SectionOption, refs }: FormSection
       notApplicable: false,
       isCustom: false,
     });
-  };
-
-  const handleIsDbGapRegisteredChange = (e, checked: boolean) => {
-    setIsdbGaPRegistered(checked);
-    if (!checked) {
-      setDbGaPPPHSNumber("");
-    }
   };
 
   /**
@@ -487,35 +479,6 @@ const FormSectionB: FC<FormSectionProps> = ({ SectionOption, refs }: FormSection
               readOnly={readOnlyInputs}
             />
           )}
-        />
-      </SectionGroup>
-
-      {/* dbGaP Registration section */}
-      <SectionGroup
-        title={SectionBMetadata.sections.DBGAP_REGISTRATION.title}
-        description={SectionBMetadata.sections.DBGAP_REGISTRATION.description}
-      >
-        <SwitchInput
-          id="section-b-dbGaP-registration"
-          label="Has your study been registered in dbGaP?"
-          name="study[isDbGapRegistered]"
-          required
-          value={isDbGapRegistered}
-          onChange={handleIsDbGapRegisteredChange}
-          isBoolean
-          readOnly={readOnlyInputs}
-        />
-        <TextInput
-          id="section-b-if-yes-provide-dbgap-phs-number"
-          label="If yes, provide dbGaP PHS number with the version number"
-          name="study[dbGaPPPHSNumber]"
-          value={dbGaPPPHSNumber}
-          onChange={(e) => setDbGaPPPHSNumber(e.target.value || "")}
-          maxLength={50}
-          placeholder={'Ex/ "phs002529.v1.p1". 50 characters allowed'}
-          gridWidth={12}
-          readOnly={readOnlyInputs || !isDbGapRegistered}
-          required={isDbGapRegistered}
         />
       </SectionGroup>
